actualizar precios y favicon

+++ app/pago/CartSummary.tsx
@@ -128,6 +128,7 @@ const CartSummary = () => {
             body: JSON.stringify({ ecommerce, eventId, userIpAddress })
         })
     }
+
     return (
         <div className='flex flex-col py-24 px-24 sm:px-64 lg:px-0 lg:border-custom lg:rounded-custom w-full lg:w-1/3 h-fit bg-light-background lg:bg-background'>
             <div className={`flex transition-all duration-300 ease-linear lg:hidden ${summaryOpened ? 'mb-24' : ''} justify-between`}>
@@ -135,7 +136,7 @@ const CartSummary = () => {
                     <h2 className='font-semibold text-primary'>Mostrar resúmen de mi pedido</h2>
                     <KeyboardArrowDown className={`text-primary ${summaryOpened ? 'rotate-180' : ''}`}></KeyboardArrowDown>
                 </div>
-                <span className='font-medium text-heading'>${parseFloat(total?.toFixed(2)).toLocaleString('es-ES', { minimumFractionDigits: 2 })}<span className='text-small text-text-faded ml-6'>USD</span></span>
+                <span className='font-medium text-heading'>${parseInt(total?.toFixed(2)).toLocaleString('es-ES', { minimumFractionDigits: 2})}<span className='text-small text-text-faded ml-6'>USD</span></span>
             </div>
             <div className={`transition-all duration-300 ease-linear ${summaryOpened ? 'max-h-[2000px]' : 'max-h-0'} overflow-hidden lg:max-h-full flex flex-col lg:space-y-24`}>
                 <h2 className='hidden lg:block font-medium text-heading leading-body pb-12 lg:mx-24 border-b-custom text-center'>Resúmen del pedido</h2>
